Refactor out repositories into their own package
