+++ src/main/java/uk/gov/mca/beacons/service/registrations/PersonRepository.java
@@ -1,8 +1,9 @@
-package uk.gov.mca.beacons.service.model;
+package uk.gov.mca.beacons.service.registrations;
 
 import java.util.UUID;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+import uk.gov.mca.beacons.service.model.Person;
 
 @Repository
 public interface PersonRepository extends CrudRepository<Person, UUID> {}
