+++ src/main/java/uk/gov/mca/beacons/service/registrations/VesselRepository.java
@@ -1,8 +1,9 @@
-package uk.gov.mca.beacons.service.model;
+package uk.gov.mca.beacons.service.registrations;
 
 import java.util.UUID;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+import uk.gov.mca.beacons.service.model.Vessel;
 
 @Repository
 public interface VesselRepository extends CrudRepository<Vessel, UUID> {}
